minor package structure rework
