+++ src/main/java/com/gottmusig/database/service/configuration/DatabaseServiceConfiguration.java
@@ -2,19 +2,19 @@ package com.gottmusig.database.service.configuration;
 
 import com.gottmusig.database.service.domain.GottMusIg;
 import com.gottmusig.database.service.domain.account.AccountService;
+import com.gottmusig.database.service.domain.account.jpa.AccountServiceImpl;
 import com.gottmusig.database.service.domain.character.CharacterService;
+import com.gottmusig.database.service.domain.character.jpa.CharacterServiceImpl;
+import com.gottmusig.database.service.domain.character.jpa.blizzard.SearchCharacterClient;
 import com.gottmusig.database.service.domain.dpsdifference.DPSDifferenceService;
+import com.gottmusig.database.service.domain.dpsdifference.jpa.DPSDifferenceServiceImpl;
 import com.gottmusig.database.service.domain.item.ItemService;
+import com.gottmusig.database.service.domain.item.jpa.ItemServiceImpl;
+import com.gottmusig.database.service.domain.jpa.GottMusIgImpl;
 import com.gottmusig.database.service.domain.realm.RealmService;
+import com.gottmusig.database.service.domain.realm.jpa.RealmServiceImpl;
 import com.gottmusig.database.service.domain.simulation.SimulationService;
-import com.gottmusig.database.service.jpa.GottMusIgImpl;
-import com.gottmusig.database.service.jpa.account.AccountServiceImpl;
-import com.gottmusig.database.service.jpa.character.CharacterServiceImpl;
-import com.gottmusig.database.service.jpa.character.blizzard.SearchCharacterClient;
-import com.gottmusig.database.service.jpa.dpsdifference.DPSDifferenceServiceImpl;
-import com.gottmusig.database.service.jpa.item.ItemServiceImpl;
-import com.gottmusig.database.service.jpa.realm.RealmServiceImpl;
-import com.gottmusig.database.service.jpa.simulation.SimulationServiceImpl;
+import com.gottmusig.database.service.domain.simulation.jpa.SimulationServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
