+++ src/main/java/com/gottmusig/database/service/domain/character/jpa/characterpojo/Spec.java
@@ -1,0 +1,72 @@
+package com.gottmusig.database.service.domain.character.jpa.characterpojo;
+
+import com.google.gson.annotations.SerializedName;
+
+import javax.annotation.Generated;
+
+@Generated("net.hexar.json2pojo")
+@SuppressWarnings("unused")
+public class Spec {
+
+    @SerializedName("backgroundImage")
+    private String mBackgroundImage;
+    @SerializedName("description")
+    private String mDescription;
+    @SerializedName("icon")
+    private String mIcon;
+    @SerializedName("name")
+    private String mName;
+    @SerializedName("order")
+    private Long mOrder;
+    @SerializedName("role")
+    private String mRole;
+
+    public String getBackgroundImage() {
+        return mBackgroundImage;
+    }
+
+    public void setBackgroundImage(String backgroundImage) {
+        mBackgroundImage = backgroundImage;
+    }
+
+    public String getDescription() {
+        return mDescription;
+    }
+
+    public void setDescription(String description) {
+        mDescription = description;
+    }
+
+    public String getIcon() {
+        return mIcon;
+    }
+
+    public void setIcon(String icon) {
+        mIcon = icon;
+    }
+
+    public String getName() {
+        return mName;
+    }
+
+    public void setName(String name) {
+        mName = name;
+    }
+
+    public Long getOrder() {
+        return mOrder;
+    }
+
+    public void setOrder(Long order) {
+        mOrder = order;
+    }
+
+    public String getRole() {
+        return mRole;
+    }
+
+    public void setRole(String role) {
+        mRole = role;
+    }
+
+}
