+++ src/main/java/com/gottmusig/database/service/domain/item/jpa/ItemServiceImpl.java
@@ -1,0 +1,14 @@
+package com.gottmusig.database.service.domain.item.jpa;
+
+import com.gottmusig.database.service.domain.item.ItemService;
+import org.springframework.stereotype.Service;
+
+/**
+ * Description
+ *
+ * @author lgottschick
+ * @since 1.0.0-SNAPSHOT
+ */
+@Service
+public class ItemServiceImpl implements ItemService {
+}
