+++ src/main/java/com/gottmusig/database/service/domain/simulation/jpa/SimulationServiceImpl.java
@@ -1,0 +1,20 @@
+package com.gottmusig.database.service.domain.simulation.jpa;
+
+import com.gottmusig.database.service.domain.character.Character;
+import com.gottmusig.database.service.domain.simulation.SimulationService;
+import org.springframework.stereotype.Service;
+
+/**
+ * Description
+ *
+ * @author lgottschick
+ * @since 1.0.0-SNAPSHOT
+ */
+@Service
+public class SimulationServiceImpl implements SimulationService {
+
+    @Override
+    public Character simulateDPS(Character character) {
+        return null;
+    }
+}
