--- src/main/java/com/gottmusig/database/service/jpa/character/WOWClassEntity.java
@@ -1,51 +1,0 @@
-package com.gottmusig.database.service.jpa.character;
-
-import com.gottmusig.database.service.domain.character.WOWClass;
-import com.gottmusig.database.service.jpa.NumericSequenceId;
-import org.springframework.data.repository.CrudRepository;
-
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-/**
- * @author leong
- * @since 20.11.2016
- */
-@Entity
-@Table(name = "wowclass")
-public class WOWClassEntity implements WOWClass {
-
-    @EmbeddedId
-    private NumericSequenceId id;
-
-    private String name;
-
-    public WOWClassEntity() {
-        this.id = new NumericSequenceId();
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public NumericSequenceId getId() {
-        return id;
-    }
-
-    public void setId(NumericSequenceId id) {
-        this.id = id;
-    }
-
-    public interface WOWClassRepository extends CrudRepository<WOWClassEntity, NumericSequenceId> {
-
-        WOWClassEntity findByName(String name);
-
-    }
-}
