--- src/main/java/com/gottmusig/database/service/jpa/item/ItemServiceImpl.java
@@ -1,14 +1,0 @@
-package com.gottmusig.database.service.jpa.item;
-
-import com.gottmusig.database.service.domain.item.ItemService;
-import org.springframework.stereotype.Service;
-
-/**
- * Description
- *
- * @author lgottschick
- * @since 1.0.0-SNAPSHOT
- */
-@Service
-public class ItemServiceImpl implements ItemService {
-}
